Added the slice command to isa-clips
The slice command takes a 64-bit number and takes a slice of the size specified
by a start and end location within the number. So (slice ? 0 41) will return
the lower 41-bits of a 64-bit number.

Currently, I just use a long long to store the given values but I make an
assumption that long longs are 64-bits in length. Eventually I will change this
to use the sizeof(long long) instead.

Internally the slice command uses unsigned long longs but returns signed ones
so that clips doesn't filp out if it turns out that unsigned value is larger
than the signed long long clips uses internally.

Using unsigned long long also ensures that slice does a logical instead of an
arithmetic shift.