Added slice8, slice4, and slice2
These functions will take in a 64-bit number and divide it into 8, 4, or 2
slices.

slice2 returns two 32-bit numbers representing the given 64-bit number
slice4 returns four 16-bit numbers representing the given 64-bit number
slice8 returns eight 8-bit numbers representing the given 64-bit number

The ordering within the multifield is least significant to most significant. So
for slice8 the ordering is as follows
( (0:8) (8:16) (16:24) (24:32) (32:40) (40:48) (48:56) (56:64) )
for slice4
( (0:16) (16:32) (32:48) (48:64) )
for slice2
( (0:32) (32:64) )

If you give slice2,4,8 too many arguments or something that isn't an integer
then the result will be a multifield of length 0.